requirements.txt
matplotlib
pillow
pydub
numpy
tqdm
